+++ script/tool/test/xcode_analyze_command_test.dart
@@ -2,7 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:convert';
 import 'dart:io' as io;
 
 import 'package:args/command_runner.dart';
@@ -10,50 +9,16 @@ import 'package:file/file.dart';
 import 'package:file/memory.dart';
 import 'package:flutter_plugin_tools/src/common/core.dart';
 import 'package:flutter_plugin_tools/src/common/plugin_utils.dart';
-import 'package:flutter_plugin_tools/src/xctest_command.dart';
+import 'package:flutter_plugin_tools/src/xcode_analyze_command.dart';
 import 'package:test/test.dart';
 
 import 'mocks.dart';
 import 'util.dart';
 
-final Map<String, dynamic> _kDeviceListMap = <String, dynamic>{
-  'runtimes': <Map<String, dynamic>>[
-    <String, dynamic>{
-      'bundlePath':
-          '/Library/Developer/CoreSimulator/Profiles/Runtimes/iOS 13.4.simruntime',
-      'buildversion': '17L255',
-      'runtimeRoot':
-          '/Library/Developer/CoreSimulator/Profiles/Runtimes/iOS 13.4.simruntime/Contents/Resources/RuntimeRoot',
-      'identifier': 'com.apple.CoreSimulator.SimRuntime.iOS-13-4',
-      'version': '13.4',
-      'isAvailable': true,
-      'name': 'iOS 13.4'
-    },
-  ],
-  'devices': <String, dynamic>{
-    'com.apple.CoreSimulator.SimRuntime.iOS-13-4': <Map<String, dynamic>>[
-      <String, dynamic>{
-        'dataPath':
-            '/Users/xxx/Library/Developer/CoreSimulator/Devices/1E76A0FD-38AC-4537-A989-EA639D7D012A/data',
-        'logPath':
-            '/Users/xxx/Library/Logs/CoreSimulator/1E76A0FD-38AC-4537-A989-EA639D7D012A',
-        'udid': '1E76A0FD-38AC-4537-A989-EA639D7D012A',
-        'isAvailable': true,
-        'deviceTypeIdentifier':
-            'com.apple.CoreSimulator.SimDeviceType.iPhone-8-Plus',
-        'state': 'Shutdown',
-        'name': 'iPhone 8 Plus'
-      }
-    ]
-  }
-};
-
 // TODO(stuartmorgan): Rework these tests to use a mock Xcode instead of
 // doing all the process mocking and validation.
 void main() {
-  const String _kDestination = '--ios-destination';
-
-  group('test xctest_command', () {
+  group('test xcode_analyze_command', () {
     late FileSystem fileSystem;
     late MockPlatform mockPlatform;
     late Directory packagesDir;
@@ -65,17 +30,18 @@ void main() {
       mockPlatform = MockPlatform(isMacOS: true);
       packagesDir = createPackagesDirectory(fileSystem: fileSystem);
       processRunner = RecordingProcessRunner();
-      final XCTestCommand command = XCTestCommand(packagesDir,
+      final XcodeAnalyzeCommand command = XcodeAnalyzeCommand(packagesDir,
           processRunner: processRunner, platform: mockPlatform);
 
-      runner = CommandRunner<void>('xctest_command', 'Test for xctest_command');
+      runner = CommandRunner<void>(
+          'xcode_analyze_command', 'Test for xcode_analyze_command');
       runner.addCommand(command);
     });
 
     test('Fails if no platforms are provided', () async {
       Error? commandError;
       final List<String> output = await runCapturingPrint(
-          runner, <String>['xctest'], errorHandler: (Error e) {
+          runner, <String>['xcode-analyze'], errorHandler: (Error e) {
         commandError = e;
       });
 
@@ -88,94 +54,6 @@ void main() {
       );
     });
 
-    test('allows target filtering', () async {
-      final Directory pluginDirectory1 = createFakePlugin('plugin', packagesDir,
-          platformSupport: <String, PlatformSupport>{
-            kPlatformMacos: PlatformSupport.inline,
-          });
-
-      final Directory pluginExampleDirectory =
-          pluginDirectory1.childDirectory('example');
-
-      processRunner.processToReturn = MockProcess.succeeding();
-      processRunner.resultStdout =
-          '{"project":{"targets":["bar_scheme", "foo_scheme"]}}';
-      final List<String> output = await runCapturingPrint(runner, <String>[
-        'xctest',
-        '--macos',
-        '--test-target=RunnerTests',
-      ]);
-
-      expect(
-          output,
-          contains(
-              contains('Successfully ran macOS xctest for plugin/example')));
-
-      expect(
-          processRunner.recordedCalls,
-          orderedEquals(<ProcessCall>[
-            ProcessCall(
-                'xcrun',
-                const <String>[
-                  'xcodebuild',
-                  'test',
-                  '-workspace',
-                  'macos/Runner.xcworkspace',
-                  '-scheme',
-                  'Runner',
-                  '-configuration',
-                  'Debug',
-                  '-only-testing:RunnerTests',
-                  'GCC_TREAT_WARNINGS_AS_ERRORS=YES',
-                ],
-                pluginExampleDirectory.path),
-          ]));
-    });
-
-    test('reports skips with no tests', () async {
-      final Directory pluginDirectory1 = createFakePlugin('plugin', packagesDir,
-          platformSupport: <String, PlatformSupport>{
-            kPlatformMacos: PlatformSupport.inline,
-          });
-
-      final Directory pluginExampleDirectory =
-          pluginDirectory1.childDirectory('example');
-
-      // Exit code 66 from testing indicates no tests.
-      final MockProcess noTestsProcessResult = MockProcess();
-      noTestsProcessResult.exitCodeCompleter.complete(66);
-
-      processRunner.processToReturn = MockProcess.succeeding();
-      processRunner.resultStdout =
-          '{"project":{"targets":["bar_scheme", "foo_scheme"]}}';
-      processRunner.mockProcessesForExecutable['xcrun'] = <io.Process>[
-        noTestsProcessResult,
-      ];
-      final List<String> output =
-          await runCapturingPrint(runner, <String>['xctest', '--macos']);
-
-      expect(output, contains(contains('No tests found.')));
-
-      expect(
-          processRunner.recordedCalls,
-          orderedEquals(<ProcessCall>[
-            ProcessCall(
-                'xcrun',
-                const <String>[
-                  'xcodebuild',
-                  'test',
-                  '-workspace',
-                  'macos/Runner.xcworkspace',
-                  '-scheme',
-                  'Runner',
-                  '-configuration',
-                  'Debug',
-                  'GCC_TREAT_WARNINGS_AS_ERRORS=YES',
-                ],
-                pluginExampleDirectory.path),
-          ]));
-    });
-
     group('iOS', () {
       test('skip if iOS is not supported', () async {
         createFakePlugin('plugin', packagesDir,
@@ -183,12 +61,10 @@ void main() {
               kPlatformMacos: PlatformSupport.inline,
             });
 
-        final List<String> output = await runCapturingPrint(runner,
-            <String>['xctest', '--ios', _kDestination, 'foo_destination']);
-        expect(
-            output,
-            contains(
-                contains('iOS is not implemented by this plugin package.')));
+        final List<String> output =
+            await runCapturingPrint(runner, <String>['xcode-analyze', '--ios']);
+        expect(output,
+            contains(contains('Not implemented for target platform(s).')));
         expect(processRunner.recordedCalls, orderedEquals(<ProcessCall>[]));
       });
 
@@ -198,16 +74,14 @@ void main() {
               kPlatformIos: PlatformSupport.federated
             });
 
-        final List<String> output = await runCapturingPrint(runner,
-            <String>['xctest', '--ios', _kDestination, 'foo_destination']);
-        expect(
-            output,
-            contains(
-                contains('iOS is not implemented by this plugin package.')));
+        final List<String> output =
+            await runCapturingPrint(runner, <String>['xcode-analyze', '--ios']);
+        expect(output,
+            contains(contains('Not implemented for target platform(s).')));
         expect(processRunner.recordedCalls, orderedEquals(<ProcessCall>[]));
       });
 
-      test('running with correct destination', () async {
+      test('runs for iOS plugin', () async {
         final Directory pluginDirectory = createFakePlugin(
             'plugin', packagesDir, platformSupport: <String, PlatformSupport>{
           kPlatformIos: PlatformSupport.inline
@@ -216,21 +90,16 @@ void main() {
         final Directory pluginExampleDirectory =
             pluginDirectory.childDirectory('example');
 
-        processRunner.processToReturn = MockProcess.succeeding();
-        processRunner.resultStdout =
-            '{"project":{"targets":["bar_scheme", "foo_scheme"]}}';
         final List<String> output = await runCapturingPrint(runner, <String>[
-          'xctest',
+          'xcode-analyze',
           '--ios',
-          _kDestination,
-          'foo_destination',
         ]);
 
         expect(
             output,
             containsAllInOrder(<Matcher>[
               contains('Running for plugin'),
-              contains('Successfully ran iOS xctest for plugin/example')
+              contains('plugin/example (iOS) passed analysis.')
             ]));
 
         expect(
@@ -240,7 +109,7 @@ void main() {
                   'xcrun',
                   const <String>[
                     'xcodebuild',
-                    'test',
+                    'analyze',
                     '-workspace',
                     'ios/Runner.xcworkspace',
                     '-scheme',
@@ -248,53 +117,7 @@ void main() {
                     '-configuration',
                     'Debug',
                     '-destination',
-                    'foo_destination',
-                    'GCC_TREAT_WARNINGS_AS_ERRORS=YES',
-                  ],
-                  pluginExampleDirectory.path),
-            ]));
-      });
-
-      test('Not specifying --ios-destination assigns an available simulator',
-          () async {
-        final Directory pluginDirectory = createFakePlugin(
-            'plugin', packagesDir, platformSupport: <String, PlatformSupport>{
-          kPlatformIos: PlatformSupport.inline
-        });
-
-        final Directory pluginExampleDirectory =
-            pluginDirectory.childDirectory('example');
-
-        final Map<String, dynamic> schemeCommandResult = <String, dynamic>{
-          'project': <String, dynamic>{
-            'targets': <String>['bar_scheme', 'foo_scheme']
-          }
-        };
-        processRunner.processToReturn = MockProcess.succeeding();
-        // For simplicity of the test, we combine all the mock results into a single mock result, each internal command
-        // will get this result and they should still be able to parse them correctly.
-        processRunner.resultStdout =
-            jsonEncode(schemeCommandResult..addAll(_kDeviceListMap));
-        await runCapturingPrint(runner, <String>['xctest', '--ios']);
-
-        expect(
-            processRunner.recordedCalls,
-            orderedEquals(<ProcessCall>[
-              const ProcessCall(
-                  'xcrun', <String>['simctl', 'list', '--json'], null),
-              ProcessCall(
-                  'xcrun',
-                  const <String>[
-                    'xcodebuild',
-                    'test',
-                    '-workspace',
-                    'ios/Runner.xcworkspace',
-                    '-scheme',
-                    'Runner',
-                    '-configuration',
-                    'Debug',
-                    '-destination',
-                    'id=1E76A0FD-38AC-4537-A989-EA639D7D012A',
+                    'generic/platform=iOS Simulator',
                     'GCC_TREAT_WARNINGS_AS_ERRORS=YES',
                   ],
                   pluginExampleDirectory.path),
@@ -307,9 +130,6 @@ void main() {
               kPlatformIos: PlatformSupport.inline
             });
 
-        processRunner.processToReturn = MockProcess.succeeding();
-        processRunner.resultStdout =
-            '{"project":{"targets":["bar_scheme", "foo_scheme"]}}';
         processRunner.mockProcessesForExecutable['xcrun'] = <io.Process>[
           MockProcess.failing()
         ];
@@ -318,10 +138,8 @@ void main() {
         final List<String> output = await runCapturingPrint(
           runner,
           <String>[
-            'xctest',
+            'xcode-analyze',
             '--ios',
-            _kDestination,
-            'foo_destination',
           ],
           errorHandler: (Error e) {
             commandError = e;
@@ -340,14 +158,15 @@ void main() {
 
     group('macOS', () {
       test('skip if macOS is not supported', () async {
-        createFakePlugin('plugin', packagesDir);
+        createFakePlugin(
+          'plugin',
+          packagesDir,
+        );
 
-        final List<String> output =
-            await runCapturingPrint(runner, <String>['xctest', '--macos']);
-        expect(
-            output,
-            contains(
-                contains('macOS is not implemented by this plugin package.')));
+        final List<String> output = await runCapturingPrint(
+            runner, <String>['xcode-analyze', '--macos']);
+        expect(output,
+            contains(contains('Not implemented for target platform(s).')));
         expect(processRunner.recordedCalls, orderedEquals(<ProcessCall>[]));
       });
 
@@ -357,12 +176,10 @@ void main() {
               kPlatformMacos: PlatformSupport.federated,
             });
 
-        final List<String> output =
-            await runCapturingPrint(runner, <String>['xctest', '--macos']);
-        expect(
-            output,
-            contains(
-                contains('macOS is not implemented by this plugin package.')));
+        final List<String> output = await runCapturingPrint(
+            runner, <String>['xcode-analyze', '--macos']);
+        expect(output,
+            contains(contains('Not implemented for target platform(s).')));
         expect(processRunner.recordedCalls, orderedEquals(<ProcessCall>[]));
       });
 
@@ -376,18 +193,13 @@ void main() {
         final Directory pluginExampleDirectory =
             pluginDirectory1.childDirectory('example');
 
-        processRunner.processToReturn = MockProcess.succeeding();
-        processRunner.resultStdout =
-            '{"project":{"targets":["bar_scheme", "foo_scheme"]}}';
         final List<String> output = await runCapturingPrint(runner, <String>[
-          'xctest',
+          'xcode-analyze',
           '--macos',
         ]);
 
-        expect(
-            output,
-            contains(
-                contains('Successfully ran macOS xctest for plugin/example')));
+        expect(output,
+            contains(contains('plugin/example (macOS) passed analysis.')));
 
         expect(
             processRunner.recordedCalls,
@@ -396,7 +208,7 @@ void main() {
                   'xcrun',
                   const <String>[
                     'xcodebuild',
-                    'test',
+                    'analyze',
                     '-workspace',
                     'macos/Runner.xcworkspace',
                     '-scheme',
@@ -415,16 +227,14 @@ void main() {
               kPlatformMacos: PlatformSupport.inline,
             });
 
-        processRunner.processToReturn = MockProcess.succeeding();
-        processRunner.resultStdout =
-            '{"project":{"targets":["bar_scheme", "foo_scheme"]}}';
         processRunner.mockProcessesForExecutable['xcrun'] = <io.Process>[
           MockProcess.failing()
         ];
 
         Error? commandError;
         final List<String> output = await runCapturingPrint(
-            runner, <String>['xctest', '--macos'], errorHandler: (Error e) {
+            runner, <String>['xcode-analyze', '--macos'],
+            errorHandler: (Error e) {
           commandError = e;
         });
 
@@ -451,22 +261,17 @@ void main() {
         final Directory pluginExampleDirectory =
             pluginDirectory1.childDirectory('example');
 
-        processRunner.processToReturn = MockProcess.succeeding();
-        processRunner.resultStdout =
-            '{"project":{"targets":["bar_scheme", "foo_scheme"]}}';
         final List<String> output = await runCapturingPrint(runner, <String>[
-          'xctest',
+          'xcode-analyze',
           '--ios',
           '--macos',
-          _kDestination,
-          'foo_destination',
         ]);
 
         expect(
             output,
             containsAll(<Matcher>[
-              contains('Successfully ran iOS xctest for plugin/example'),
-              contains('Successfully ran macOS xctest for plugin/example'),
+              contains('plugin/example (iOS) passed analysis.'),
+              contains('plugin/example (macOS) passed analysis.'),
             ]));
 
         expect(
@@ -476,7 +281,7 @@ void main() {
                   'xcrun',
                   const <String>[
                     'xcodebuild',
-                    'test',
+                    'analyze',
                     '-workspace',
                     'ios/Runner.xcworkspace',
                     '-scheme',
@@ -484,7 +289,7 @@ void main() {
                     '-configuration',
                     'Debug',
                     '-destination',
-                    'foo_destination',
+                    'generic/platform=iOS Simulator',
                     'GCC_TREAT_WARNINGS_AS_ERRORS=YES',
                   ],
                   pluginExampleDirectory.path),
@@ -492,7 +297,7 @@ void main() {
                   'xcrun',
                   const <String>[
                     'xcodebuild',
-                    'test',
+                    'analyze',
                     '-workspace',
                     'macos/Runner.xcworkspace',
                     '-scheme',
@@ -515,22 +320,16 @@ void main() {
         final Directory pluginExampleDirectory =
             pluginDirectory1.childDirectory('example');
 
-        processRunner.processToReturn = MockProcess.succeeding();
-        processRunner.resultStdout =
-            '{"project":{"targets":["bar_scheme", "foo_scheme"]}}';
         final List<String> output = await runCapturingPrint(runner, <String>[
-          'xctest',
+          'xcode-analyze',
           '--ios',
           '--macos',
-          _kDestination,
-          'foo_destination',
         ]);
 
         expect(
             output,
             containsAllInOrder(<Matcher>[
-              contains('Only running for macOS'),
-              contains('Successfully ran macOS xctest for plugin/example'),
+              contains('plugin/example (macOS) passed analysis.'),
             ]));
 
         expect(
@@ -540,7 +339,7 @@ void main() {
                   'xcrun',
                   const <String>[
                     'xcodebuild',
-                    'test',
+                    'analyze',
                     '-workspace',
                     'macos/Runner.xcworkspace',
                     '-scheme',
@@ -562,23 +361,16 @@ void main() {
         final Directory pluginExampleDirectory =
             pluginDirectory.childDirectory('example');
 
-        processRunner.processToReturn = MockProcess.succeeding();
-        processRunner.resultStdout =
-            '{"project":{"targets":["bar_scheme", "foo_scheme"]}}';
         final List<String> output = await runCapturingPrint(runner, <String>[
-          'xctest',
+          'xcode-analyze',
           '--ios',
           '--macos',
-          _kDestination,
-          'foo_destination',
         ]);
 
         expect(
             output,
-            containsAllInOrder(<Matcher>[
-              contains('Only running for iOS'),
-              contains('Successfully ran iOS xctest for plugin/example')
-            ]));
+            containsAllInOrder(
+                <Matcher>[contains('plugin/example (iOS) passed analysis.')]));
 
         expect(
             processRunner.recordedCalls,
@@ -587,7 +379,7 @@ void main() {
                   'xcrun',
                   const <String>[
                     'xcodebuild',
-                    'test',
+                    'analyze',
                     '-workspace',
                     'ios/Runner.xcworkspace',
                     '-scheme',
@@ -595,7 +387,7 @@ void main() {
                     '-configuration',
                     'Debug',
                     '-destination',
-                    'foo_destination',
+                    'generic/platform=iOS Simulator',
                     'GCC_TREAT_WARNINGS_AS_ERRORS=YES',
                   ],
                   pluginExampleDirectory.path),
@@ -605,22 +397,16 @@ void main() {
       test('skips when neither are supported', () async {
         createFakePlugin('plugin', packagesDir);
 
-        processRunner.processToReturn = MockProcess.succeeding();
-        processRunner.resultStdout =
-            '{"project":{"targets":["bar_scheme", "foo_scheme"]}}';
         final List<String> output = await runCapturingPrint(runner, <String>[
-          'xctest',
+          'xcode-analyze',
           '--ios',
           '--macos',
-          _kDestination,
-          'foo_destination',
         ]);
 
         expect(
             output,
             containsAllInOrder(<Matcher>[
-              contains(
-                  'SKIPPING: Neither iOS nor macOS is implemented by this plugin package.'),
+              contains('SKIPPING: Not implemented for target platform(s).'),
             ]));
 
         expect(processRunner.recordedCalls, orderedEquals(<ProcessCall>[]));