+++ script/tool/CHANGELOG.md
@@ -1,6 +1,9 @@
 ## NEXT
 
 - Improved `license-check` output.
+- Added an `xctest` flag to select specific test targets, to allow running only
+  unit tests or integration tests.
+- Split Xcode analysis out of `xctest` and into a new `xcode-analyze` command.
 
 ## 0.4.0